perf(leaderboard): instant weekly tab switches + parallel backend queries
- Lift the weekly fetch into LeaderboardTabs so toggling tabs no longer
  unmounts WeeklyTab and re-fires the request. Subsequent switches are
  instant; the loading state only appears on first paint.
- Run the top-user details fetch and the global ranking fetch in
  parallel via Promise.all in /api/leaderboard?range=week. Saves one
  Supabase round-trip per request.
- Add max-age=30 to Cache-Control so browsers cache the response too
  (previously only the Vercel CDN cached via s-maxage). Quick re-visits
  within 30s skip the network entirely.
